Centralize font code to RenderBuffer
